--- src/main/java/by/fiveelement/MainClass.java
@@ -1,4 +1,0 @@
-package by.fiveelement;
-
-public class MainClass {
-}
