+++ src/main/java/by/fiveelement/ui/page/cart/CartPageMessage.java
@@ -1,4 +1,4 @@
-package by.fiveelement;
+package by.fiveelement.ui.page.cart;
 
 public class CartPageMessage {
     public static final String TEXT_ADD_TO_CART="Товар добавлен в корзину";
