+++ src/main/java/by/fiveelement/ui/page/registration/RegistrationMessage.java
@@ -1,4 +1,4 @@
-package by.fiveelement;
+package by.fiveelement.ui.page.registration;
 
 public class RegistrationMessage {
