+++ src/main/java/by/fiveelement/util/GenerationDataUtil.java
