--- src/test/java/by/fiveelement/TestClass.java
@@ -1,5 +1,0 @@
-package by.fiveelement;
-
-public class TestClass {
-
-}
