+++ src/test/java/by/fiveelement/api/ForgotPasswordApiTest.java
@@ -1,5 +1,6 @@
-package by.fiveelement;
+package by.fiveelement.api;
 
+import by.fiveelement.ForgotPasswordApiRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
