+++ src/test/java/by/fiveelement/api/UserRegistrationApiTest.java
@@ -1,4 +1,4 @@
-package by.fiveelement;
+package by.fiveelement.api;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
